Code cleanup, moved query items to HTTPLabelSerializer
